Move diffing code back to common
