+++ cli/common/diff_test.go
@@ -1,4 +1,4 @@
-package helm
+package common
 
 import (
 	"testing"
